+++ docs/src/content/features/_meta.ts
@@ -3,6 +3,7 @@ export default {
   'card-farming': 'Card Farming',
   'achievement-unlocker': 'Achievement Unlocker',
   'achievement-manager': 'Achievement Manager',
+  'trading-card-manager': 'Trading Card Manager',
   'playtime-booster': 'Playtime Booster',
   'auto-idler': 'Automatic Idler',
 }